Move useTheme custom hook
Move useTheme into a separate useTheme.js file to fix the warning that fast refresh wouldn't work with the hook in the ThemeContext.jsx file.
