+++ src/context/ThemeContext.jsx
@@ -1,4 +1,4 @@
-import { createContext, useContext, useEffect, useState } from 'react';
+import { createContext, useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 
 export const ThemeContext = createContext();
@@ -21,15 +21,6 @@ const ThemeProvider = ({ children }) => {
       {children}
     </ThemeContext.Provider>
   );
-};
-
-// Custom hook to throw an error if theme is undefined
-export const useTheme = () => {
-  const context = useContext(ThemeContext);
-  if (context === undefined) {
-    throw new Error('Context is undefined');
-  }
-  return context;
 };
 
 ThemeProvider.propTypes = {
