+++ src/pages/WelcomePage.jsx
@@ -1,4 +1,4 @@
-import { useTheme } from '../context/ThemeContext';
+import useTheme from '../context/useTheme';
 import PrimaryLinkButton from '../components/PrimaryLinkButton';
 
 export default function Welcome() {
